add hover state to hex-content
